test: remove broken tests
It appears that it is not possible to test changes in UI on styled components that are changed due to props that are changed up the tree. Since our FormFieldLabel changes its compact appearance due to a change to the size prop on the parent (and not directly on itself), the tests does not reflect this change, and report the label to be of normal size. If the "small" prop is set on the label directly, the tests pass. I don't think we should follow this pattern just because our testing library is so restrictive, so I've decided to reomve the tests instead.