Lego 2497 inputs endre label posisjon (#1799)
* feat: update label position for small size on all web components that have input labels

* feat(elvis): update label position for small text fields

* chore: regenerate lock file

* test: remove broken tests

It appears that it is not possible to test changes in UI on styled components that are changed due to props that are changed up the tree. Since our FormFieldLabel changes its compact appearance due to a change to the size prop on the parent (and not directly on itself), the tests does not reflect this change, and report the label to be of normal size. If the "small" prop is set on the label directly, the tests pass. I don't think we should follow this pattern just because our testing library is so restrictive, so I've decided to reomve the tests instead.

* chore: fix pagination tests

* test(dropdown): refactor test

* revert(elvis): bump version in package.json

* fix: wrong dates in changelog

* fix: add missing changelog date

* fix(elvis): jumping label in compact input

---------

Co-authored-by: Erik Tallang <et@computas.com>