Lego 3647 komponenter stotte for custom aria labels pa desktopmenu mobilemenu (#3285)
* Add custom aria labels on elvia-header

Added automatic language switch support based on lang attribute on header component.
desktopMenu and mobileMenu gets their translations from labels list in elvia-header.tsx passed as props.

* Add custom aria label support on themepicker

Added more automatic label switching support on themepicker in header component. Passed lang as a prop to themepicker to decide which labels to use.

* Refactor themepicker

Rafactored themepicker to use the hook directly instead of passing it as a prop as it is not used in the files connecting the header component and themepicker.

* Remove lang as prop from header component

Removed lang as a prop from being passed to mobile and desktop menu. Code cleanup.

* Update changelog and packages

Updated changelog and packages.
Added custom aria label switch on theme label.

* Update themePicker.tsx

* Update yarn.lock

* Update packages/components/components/elvis-header/CHANGELOG.json

Co-authored-by: Tom Schrier <54577949+TomSchrier@users.noreply.github.com>

* Update date in Changelog

* Resolve merge conflict on changelog and update package

---------

Co-authored-by: Tom Schrier <54577949+TomSchrier@users.noreply.github.com>